revised project-selection and card css started context-bar primitive
