+++ src/App.scss
@@ -15,6 +15,8 @@
 @import './components/project-form/project-form';
 @import './components/project-manage/project-manage';
 @import './components/project-gallery-manage/project-gallery-manage';
+@import './components/project-selection/project-selection';
+@import './components/project-card/project-card';
 
 @import './components/artwork-form/artwork-form';
 @import './components/artwork-manage/artwork-manage';
