+++ src/components/project-card/_project-card.scss
@@ -1,0 +1,21 @@
+@import '../../stylesheets/variables';
+
+.project-card{
+    display: flex;
+    height: 100px;
+    justify-content: flex-start;
+    background-color: $color-background-white;
+    div.project-description{
+        display: flex;
+        justify-content: space-around;
+        padding: 1rem;
+        width:100%;
+        div.name{
+            font-size: 1.25rem;
+        }
+    }
+    div.project-image{
+        height:inherit;
+        border-right: 1px solid $color-primary-grey;
+    }
+}
