+++ src/components/project-manage/project-manage.component.jsx
@@ -2,23 +2,30 @@ import React from 'react';
 
 import {Route, Link} from 'react-router-dom'
 import ProjectForm from '../project-form/project-form.component';
-import SideMenu from '../side-menu.component';
+import ProjectSelection from '../project-selection/project-selection.component';
 
 class ProjectManage extends React.Component {
     constructor(){
         super();
-        this.state = {}
+        this.state = {
+            selection: true
+        }
     }
 
     componentDidMount(){
         this.props.getContext('Manage Projects');
     }
 
+
     render(){
         return(
             <div className="project-manage-container">
                 <div className="layout-container">
-                    <ProjectForm/>
+                    {(this.state.selection)?(
+                        <ProjectSelection/>
+                    ):(
+                        <ProjectForm/>
+                    )}
                 </div>
             </div>
         )
