+++ src/components/project-selection/project-selection.component.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import ProjectCard from './project-card.component';
-import ProjectService from '../services/project.service';
+import ProjectCard from '../project-card/project-card.component';
+import ProjectService from '../../services/project.service';
 
 const project$ = new ProjectService();
 
@@ -24,26 +24,34 @@ class ProjectSelection extends React.Component {
         });
     }
 
+    handleProjectClick = (id) => {
+        this.setState({selectedProject: id});
+    }
+
     projects = () => {
         const {projects, selectedProject} = this.state;
+
         return projects.map((project) => {
-           return( 
+            let className = (selectedProject  == project._id) ? 'selected' : '';
+            return( 
                 <li
                     id={project._id}
                     key={project._id}
+                    onClick={()=>{this.handleProjectClick(project._id)}}
+                    className={className}
                 >
                     <ProjectCard
                         project={project}
                     />
                 </li>
-           )
+            )
         })
     }
 
     render(){
         return(
-            <div>
-                <ul className="flex-list-y">
+            <div className="project-selection">
+                <ul>
                     {this.projects()}
                 </ul>
             </div>
